+++ src/extractSQL.ts
@@ -18,30 +18,76 @@ export async function extractSQL() {
   var queries = [];
   //loop though all lines of code
   while (i < codeLine.length) {
-    var trimmedCode = codeLine[i].toLowerCase().trim();
-    //load every text between f""" and """ into queries array.
-    //if code line ends in f""" then treat it as the start of a query
-    if (trimmedCode.substr(trimmedCode.length - 3) === "\"\"\"") {
-      var queryName = codeLine[i].trim().substr(0, codeLine[i].trim().indexOf("="));
+    //var trimmedCode = codeLine[i].toLowerCase().trim();
+    
+    //load every text between """ and """ into queries array.
+    //if code line ends in """ then treat it as the start of a query
+    var strStart = codeLine[i].indexOf("\"\"\"")
+
+    if (strStart > -1) {
+      var queryName
+      try {
+        queryName = codeLine[i].trim().substring(0, codeLine[i].trim().indexOf("="));
+        if (queryName.length == 0) {
+          throw "No query name.";
+        }
+      } catch (error) {
+        queryName = "Unnamed query"
+      }      
       queries.push(new Query(queryName));
-      i++;
-      var loopThroughQuery = true;
-      while (loopThroughQuery) {
-        queries[queries.length - 1].queryCode.push(new QueryCodeLine(codeLine[i] + '\n'));
-        if (codeLine[i].toLowerCase().includes("select")) {
+
+      var strEnd = codeLine[i].indexOf("\"\"\"", (strStart + 3))
+
+      if (strEnd > -1) {
+        //single line query- get text between """ and """ on same line
+        queries[queries.length - 1].queryCode.push(new QueryCodeLine(codeLine[i].substring((strStart+3), strEnd) + '\n'));
+        if (codeLine[i].substring((strStart+3), strEnd).toLowerCase().includes("select")) {
           queries[queries.length - 1].selectCount++;
         }
-        i++;
-        if (codeLine[i].trim() === "\"\"\"" || (i - 1) >= codeLine.length) {
-          //add all subsequent lines to query array until """ appear in a line to mark the end of the query
-          loopThroughQuery = false;
+      }
+      else {
+        //test if query starts on first line
+        if (codeLine[i].substr((strStart+3)).trim().length > 0) {
+          queries[queries.length - 1].queryCode.push(new QueryCodeLine(codeLine[i].substr((strStart+3)) + '\n'));
+          if (codeLine[i].substr((strStart+3)).toLowerCase().includes("select")) {
+            queries[queries.length - 1].selectCount++;
+          }
+        }
+        i++
+
+        var loopThroughQuery = true;
+        while (loopThroughQuery) {
+          var strEnd = codeLine[i].indexOf("\"\"\"")
+          if (strEnd > -1 || (i - 1) >= codeLine.length) {
+            //add all subsequent lines to query array until """ appear in a line to mark the end of the query
+            if (codeLine[i].substring(0, strEnd).trim().length > 0) {
+              queries[queries.length - 1].queryCode.push(new QueryCodeLine(codeLine[i].substring(0, (strEnd)) + '\n'));
+              if (codeLine[i].substring(0, (strEnd)).toLowerCase().includes("select")) {
+                queries[queries.length - 1].selectCount++;
+              }
+            }
+            loopThroughQuery = false;
+          }
+          else {
+            queries[queries.length - 1].queryCode.push(new QueryCodeLine(codeLine[i] + '\n'));
+            if (codeLine[i].toLowerCase().includes("select")) {
+              queries[queries.length - 1].selectCount++;
+            }
+            i++;
+          }
         }
       }
     }
     i++;
   }
+  console.log('start queries');
+  console.log(queries);
+  console.log('end queries');
+
   //remove any array element that doesn't have the word select in it
   var validQueries = removeInvalidQueries(queries);
+
+  console.log(validQueries);
   //number selections using the array index. Input box only seems to return selected text
   //could duplicate if queries have the same variable name
   var numberedQueries = numberQueryNames(validQueries);